% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{ANNotate}
\alias{ANNotate}
\title{Run Pan-Human Azimuth annotation}
\usage{
ANNotate(
  query_obj,
  feature_names_col = NULL,
  annotation_pipeline = "supervised",
  eval_batch_size = 40960,
  normalization_override = FALSE,
  norm_check_batch_size = 1000,
  output_mode = "minimal",
  refine_labels = TRUE,
  extract_embeddings = TRUE,
  umap_embeddings = TRUE,
  n_neighbors = 30,
  n_components = 2,
  metric = "cosine",
  min_dist = 0.3,
  umap_lr = 1,
  umap_seed = 42,
  spread = 1,
  verbose = TRUE,
  model_version = "v0",
  init = "spectral",
  process_obj = TRUE,
  cutoff_abs = 5,
  cutoff_frac = 0.001
)
}
\arguments{
\item{query_obj}{Seurat object to annotate}

\item{feature_names_col}{Column name for feature names}

\item{annotation_pipeline}{Set to 'supervised' as a default}

\item{eval_batch_size}{Evaluation batch size}

\item{normalization_override}{Whether to override normalization}

\item{norm_check_batch_size}{Batch size to inspect normalization of data}

\item{output_mode}{Output mode for annotated cell metadata}

\item{refine_labels}{Whether to refine labels}

\item{extract_embeddings}{Whether to azimuth embeddings}

\item{umap_embeddings}{Whether to include UMAP embeddings}

\item{n_neighbors}{Number of neighbors for UMAP}

\item{n_components}{Number of components for UMAP}

\item{metric}{Distance metric for UMAP}

\item{min_dist}{Minimum distance for UMAP}

\item{umap_lr}{Learning rate for UMAP}

\item{umap_seed}{Seed for UMAP}

\item{spread}{Spread parameter for UMAP}

\item{verbose}{Whether to show progress}

\item{model_version}{Version of the model to use}

\item{init}{Initialization method for UMAP}

\item{process_obj}{Whether to process the object}

\item{cutoff_abs}{Absolute cutoff for label filtering}

\item{cutoff_frac}{Fractional cutoff for label filtering}
}
\value{
Annotated Seurat object
}
\description{
Run Pan-Human Azimuth annotation
}
