% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{check_api_version}
\alias{check_api_version}
\title{Check local vs latest API version / ensure connection can be established to the server}
\usage{
check_api_version(api_base_url)
}
\arguments{
\item{api_base_url}{Base URL for the API}
}
\description{
Check local vs latest API version / ensure connection can be established to the server
}
