% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{CloudAzimuth}
\alias{CloudAzimuth}
\title{Run Pan-human Azimuth annotation on the cloud}
\usage{
CloudAzimuth(
  object = object,
  assay = "RNA",
  ip = "azimuthapi.satijalab.org",
  port = NULL,
  scheme = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat object to annotate}

\item{assay}{Name of the assay to use (default: 'RNA')}

\item{ip}{Server hostname (default: 'azimuthapi.satijalab.org')}

\item{port}{Server port (default: NULL)}

\item{scheme}{Server URL scheme (default: 'https').}

\item{...}{Additional arguments for the API to pass to the model (see ANNotate function for details)}
}
\value{
Annotated Seurat object
}
\description{
Run Pan-human Azimuth annotation on the cloud
}
\details{
Use \code{CloudAzimuth(object)} for the public cloud service. The \code{ip}, \code{port}, and \code{scheme}
arguments should not be set manually by users; the default values follow standard use.
}
\concept{annotation}
