% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{download_output}
\alias{download_output}
\title{Download output file from the API}
\usage{
download_output(url, save_path)
}
\arguments{
\item{url}{Download URL}

\item{save_path}{Path to save the downloaded file}
}
\description{
Download output file from the API
}
