% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_parser.R
\name{format_annotate_args}
\alias{format_annotate_args}
\title{Format parsed arguments for ANNotate function}
\usage{
format_annotate_args(args)
}
\arguments{
\item{args}{List of parsed arguments from parse_annotate_args()}
}
\value{
Named list of formatted arguments for ANNotate function
}
\description{
Format parsed arguments for ANNotate function
}
\examples{
\dontrun{
args <- parse_annotate_args()
formatted_args <- format_annotate_args(args)
}
}
