% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{get_data}
\alias{get_data}
\title{Get data from a Seurat object layer}
\usage{
get_data(object, assay, layer = "data")
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay name}

\item{layer}{Layer name (default: "data")}
}
\value{
Layer data
}
\description{
Get data from a Seurat object layer
}
