% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_heatmaps.R
\name{make_azimuth_QC_heatmaps}
\alias{make_azimuth_QC_heatmaps}
\title{Generate QC heatmaps for Pan-human Azimuth cell type annotations, split by broad cell type categories and further subdivided if necessary}
\usage{
make_azimuth_QC_heatmaps(
  object,
  final_name = "azimuth_fine",
  level1_name = "azimuth_broad",
  full_name = "full_hierarchical_labels",
  min.final.group = 10,
  max.ids.per.plot = 15,
  ...
)
}
\arguments{
\item{object}{Seurat object containing the metadata with Azimuth annotations}

\item{final_name}{Column name in the metadata for the final Azimuth annotations (default 'azimuth_fine')}

\item{level1_name}{Column name in the metadata for the broad Azimuth categories (default 'azimuth_broad')}

\item{full_name}{Column name in the metadata for the full hierarchical labels (default 'full_hierarchical_labels')}

\item{min.final.group}{Minimum number of cells required in a final annotation group to be included in the heatmap (default 10)}

\item{max.ids.per.plot}{Maximum number of final annotation groups to include in a single heatmap (default 15)}

\item{...}{Additional arguments to pass to the make_QC_heatmap function for customizing the heatmaps}
}
\value{
A list of ggplot objects representing the QC heatmaps for each broad cell type category, further subdivided if necessary
}
\description{
Generate QC heatmaps for Pan-human Azimuth cell type annotations, split by broad cell type categories and further subdivided if necessary
}
