% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_heatmaps.R
\name{make_QC_heatmap}
\alias{make_QC_heatmap}
\title{Generate a QC heatmap for Pan-human Azimuth cell type annotations based on the top markers for each cluster}
\usage{
make_QC_heatmap(
  seurat_obj,
  group.by = NULL,
  cells.order = NULL,
  n_downsample = 200,
  save_folder_path = NULL,
  logfc_cutoff = log(2),
  n_markers = 10,
  text.size = 5,
  text.angle = 90,
  min.size = NULL,
  max.size = NULL,
  min.pct = 0.1,
  reorder = TRUE,
  switch_id = NULL,
  identity = ""
)
}
\arguments{
\item{seurat_obj}{Seurat object to plot}

\item{group.by}{Column name in the metadata to use for grouping}

\item{cells.order}{Optional vector of cell names to specify the order of cells in the heatmap}

\item{n_downsample}{Number of cells to downsample to for each cluster (default 200)}

\item{save_folder_path}{Optional folder path to save the heatmap}

\item{logfc_cutoff}{Log fold change cutoff for selecting markers (default log(2))}

\item{n_markers}{Number of top markers to select for each cluster (default 10)}

\item{text.size}{Size of the text labels in the heatmap (default 5)}

\item{text.angle}{Angle of the text labels in the heatmap (default 90)}

\item{min.size}{Minimum number of cells in a cluster to include in the heatmap}

\item{max.size}{Maximum number of cells in a cluster to include in the heatmap}

\item{min.pct}{Minimum percentage of cells expressing a gene to consider it as a marker (default 0.1)}

\item{reorder}{Whether to reorder the clusters based on hierarchical clustering of average expression profiles (default TRUE)}

\item{switch_id}{Optional column name in the metadata to switch the Idents to before plotting}

\item{identity}{Optional string to append to the saved heatmap filename}
}
\value{
A ggplot object representing the QC heatmap
}
\description{
Generate a QC heatmap for Pan-human Azimuth cell type annotations based on the top markers for each cluster
}
