% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{package_obj}
\alias{package_obj}
\title{Package embeddings into a Seurat object}
\usage{
package_obj(
  extract_embeddings,
  embeddings_dict,
  umap_embeddings,
  umap_embeddings_dict,
  query_cells_df,
  query_obj
)
}
\arguments{
\item{extract_embeddings}{Whether embeddings are extracted}

\item{embeddings_dict}{Dictionary of embeddings}

\item{umap_embeddings}{Whether UMAP embeddings are computed}

\item{umap_embeddings_dict}{Dictionary of UMAP embeddings}

\item{query_cells_df}{Cell metadata}

\item{query_obj}{Seurat object}
}
\value{
Updated Seurat object
}
\description{
Package embeddings into a Seurat object
}
