% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_parser.R
\name{parse_annotate_args}
\alias{parse_annotate_args}
\title{Parse command line arguments for the ANNotate function}
\usage{
parse_annotate_args()
}
\value{
A list of parsed arguments
}
\description{
Parse command line arguments for the ANNotate function
}
\examples{
\dontrun{
args <- parse_annotate_args()
}
}
