% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{PrepLabel}
\alias{PrepLabel}
\title{Prepare labels for annotation}
\usage{
PrepLabel(
  object,
  label_id = "final_level_label",
  newid = "PrepLabel",
  cutid = "Other",
  cutoff = 10
)
}
\arguments{
\item{object}{Seurat object}

\item{label_id}{Column name for labels}

\item{newid}{New column name for processed labels}

\item{cutid}{Label for rejected cells}

\item{cutoff}{Minimum count threshold}
}
\value{
Updated Seurat object
}
\description{
Prepare labels for annotation
}
