% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{process_azimuth_results}
\alias{process_azimuth_results}
\title{Process results from PanAzimuth API}
\usage{
process_azimuth_results(results)
}
\arguments{
\item{results}{Raw results from the API}
}
\value{
Processed results in Seurat object format
}
\description{
Process results from PanAzimuth API
}
