% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{read_obj_min}
\alias{read_obj_min}
\title{Read and process a Seurat object}
\usage{
read_obj_min(query_obj, feature_names_col, assay_default = "RNA")
}
\arguments{
\item{query_obj}{Seurat object}

\item{feature_names_col}{Column name for feature names}

\item{assay_default}{Default assay name}
}
\value{
List containing processed data
}
\description{
Read and process a Seurat object
}
