% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_seurat_object.R
\name{save_seurat_object}
\alias{save_seurat_object}
\title{Save a Seurat object to an RDS file}
\usage{
save_seurat_object(seurat_obj, filepath)
}
\arguments{
\item{seurat_obj}{A Seurat object to save}

\item{filepath}{Character string specifying the output path}
}
\value{
Invisible filepath where the object was saved
}
\description{
Saves a Seurat object to an RDS file after applying "keto diet" to reduce size
}
