% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{ANNotate}
\alias{ANNotate}
\title{Run Pan-human Azimuth annotation locally}
\usage{
ANNotate(
  query_obj,
  feature_names_col = NULL,
  annotation_pipeline = "supervised",
  eval_batch_size = 40960L,
  normalization_override = FALSE,
  norm_check_batch_size = 1000L,
  output_mode = "minimal",
  refine_labels = TRUE,
  map_to_cl = NULL,
  include_cl_id = FALSE,
  extract_embeddings = TRUE,
  umap_embeddings = TRUE,
  n_neighbors = 30L,
  n_components = 2L,
  metric = "cosine",
  min_dist = 0.3,
  umap_lr = 1,
  umap_seed = 42L,
  spread = 1,
  verbose = TRUE,
  init = "spectral",
  model_version = "v0",
  process_obj = TRUE,
  cutoff_abs = 5,
  cutoff_frac = 0.001,
  assay = NULL
)
}
\arguments{
\item{query_obj}{Seurat object to annotate}

\item{feature_names_col}{Column name for feature names}

\item{annotation_pipeline}{Set to 'supervised' as a default}

\item{eval_batch_size}{Evaluation batch size}

\item{normalization_override}{Whether to override normalization}

\item{norm_check_batch_size}{Batch size to inspect normalization of data}

\item{output_mode}{Output mode for annotated cell metadata}

\item{refine_labels}{Whether to refine labels}

\item{map_to_cl}{One or more annotation columns to map to Cell Ontology labels}

\item{include_cl_id}{Whether to add Cell Ontology IDs to the output metadata}

\item{extract_embeddings}{Whether to extract Azimuth embeddings}

\item{umap_embeddings}{Whether to include UMAP of Azimuth embeddings}

\item{n_neighbors}{Number of neighbors for UMAP}

\item{n_components}{Number of components for UMAP}

\item{metric}{Distance metric for UMAP}

\item{min_dist}{Minimum distance for UMAP}

\item{umap_lr}{Learning rate for UMAP}

\item{umap_seed}{Seed for UMAP}

\item{spread}{Spread parameter for UMAP}

\item{verbose}{Whether to show progress}

\item{init}{Initialization method for UMAP}

\item{model_version}{Version of the model to use}

\item{process_obj}{Whether to process the object}

\item{cutoff_abs}{Absolute cutoff for label filtering}

\item{cutoff_frac}{Fractional cutoff for label filtering}

\item{assay}{Assay to use for annotation}
}
\value{
Annotated Seurat object
}
\description{
This function runs the Pan-human Azimuth model on a Seurat object to annotate cell types, via reticulate and the \code{panhumanpy} Python package. \strong{We recommend using the \code{CloudAzimuth} function, which runs cloud-based annotation, can handle large datasets, and performs robust error handling.}
}
\details{
This function requires the \code{panhumanpy} Python package to be installed and accessible via reticulate.
}
\concept{annotation}
