% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{CloudAzimuth}
\alias{CloudAzimuth}
\title{Run Pan-human Azimuth annotation on the cloud}
\usage{
CloudAzimuth(
  object = object,
  assay = "RNA",
  ip = "azimuthapi.satijalab.org",
  port = 5000,
  ...
)
}
\arguments{
\item{object}{Seurat object to annotate}

\item{assay}{Name of the assay to use (default: 'RNA')}

\item{ip}{Hostname or IP address of the cloud server (default: 'azimuthapi.satijalab.org')}

\item{port}{Port number for the API (default: 5000)}

\item{...}{Additional arguments for the API to pass to the model (see ANNotate function for details)}
}
\value{
Annotated Seurat object
}
\description{
Run Pan-human Azimuth annotation on the cloud
}
\concept{annotation}
